Implement config loading + loading check tool
We now have two configs: The oslo.config config, which keeps a reference
to our driver config, and the driver config (as yaml file) that is
currently generated from NetBox data.

A separate tool (cc-config-check) is available to emulate the driver's
loading of the config. It can be used to check config loadability before
a config is shipped out to the real driver, ideally with less
dependencies.