Make sure cc-config-check reloads config every time
Probably not relevant for real world use, but at least in the tests (and
whenever using something like this tool as a module) we want to make
sure it is not using already-present module state. The cached argument
for get_driver_config() reloads the file every time this is set to true.
As a side effect the internal fabric config is overwritten as well.