Make driver_config_path no longer required in conf
This option was required whenever config_oslo is imported by something,
also by tools or when running tests. This caused unexpected interruption
of service, therefore we're marking it as no longer required. This will
most likely not be a problem, as this parameter is used to use a big
part of the config (config_driver) and we will notice when it's gone
relatively early.