Add loadable API extensions
In the future there will be multiple functions we want to call directly,
for example for pulling current config of the driver, status for agents
or network syncs, triggering a diff or sync, or maybe reloading new
config. This commit lays the groundwork for this, as we can now call
e.g.  /v2.0/cc-fabric/status to get the stub response that extensions
have been loaded. Includes tests.