Agent Switch connection, get Arista switch status
The switch agent now has a generic Switch abstraction that can be used
by the common agent to implement certian functions. Ideally we can do
much of the work in the common agent and only have to touch the
vendor-specific agents for special stuff, but we'll see how that turns
out. For testing this functionality we have a get_switch_status() method
that can be called via API.

This commit also implements the switch interface for the Arista agent,
which is our first connection to an actual device.