Add helper to get bindinghosts from ports
For portbindings we will need to get binding_hosts from a port. Normally
this is the "host" field in the driver, but for instances created by
ironic this flag is hidden in the binding_profile under
local_link_information as a list entry as "switch_info", therefore we
need some helpers for this.