driver config: add direct_binding to hostgroup
direct_binding means that a group will be bound directly onto the
fabric, i.e. the driver will do hpb, but will bind the second level
itself. No other ml2 driver will be involved in binding this port. This
is used for baremetal-on-fabric. If direct_binding is off the driver
will bind the first level, but ignore the binding request for the second
level, leaving it up to another driver to bind it.

To reduce config size the direct_binding attribute has a variable
default: It is off by default for metagroups and on by default for
non-meta-groups. This also reflects the most common use of this flag.