Implement port bindings and config generation
This commit implements most functionalities of bind_port(), including
first and second level portbindings. It also implements a generic switch
configuration abstraction via SwitchConfigUpdate() and a reference
implementation of how this can be used to generate an Arista config
update.

SwitchConfigUpdate is used to communicate a change in config to an a
switch-agent. It supports the three operations: add, replace and remove.
It currently can manage vlans, vxlan-vlan maps and on an interface
native/trunk vlans and vlan translations.

There are still a few pieces missing, especially the port-delete calls
to decomission config/segments and proper vlan ranges inside the config.