Close sec with  exit instead of ! in Arista config
Sections don't close with !, but do with exit. In nearly any case we
have a statement that properly changes the config section beforehand,
but why take the risk (one of the contenders would've been vlan y vs
router bgp x; vlan y, but that would not have been an issue with the
current order. Still though...)