Implement ml2 port deletion logic
On port deletion we now check if a binding host is still in use on a
segment. If it's gone, we check the segment on that network is still in
use and release it, if possible. A config update is created to remove
interface config (and possibly vlan/vxlan mapping config) from affected
switches.

A segment/config cleanup will also be triggered by a port update for the
old binding host, if the update indicates a binding host change.