Transport rpc config updates as dict
If we don't serialize SwitchConfigUpdate ourselves by transporting it as
a dict and the convert it to a SwitchConfigUpdate back again in the
switch agent, then oslo.messaging will do it for us. Badly. As a list of
lists without reconverting it to a SwitchConfigUpdate. So, let's just do
this ourselves.