Move config generation into generic agent class
Part of config generation was moved from the ml2 driver into the generic
switch config model. SwitchConfigUpdateList carries a list of
SwitchConfigUpdates and can also carry out the execution. This class
will get more methods in the future to configured bindings (i.e. for
BGWs / Transits). This will then be used by the API and the syncloop as
well.