Preliminary sync network/switch API
This is a preliminary implementation of the show/sync network/switch
API. It can generate the config for a list of networks or a list of
switches that should be present on a switch and can also push said
config onto the switches.

To generate a switch config out of the blue without the top segment
being provided to us we needed to implement
get_top_level_vxlan_segments(), which will provide this information for
a list of networks.

The API will change in the future, both in layout and how we are
handling changing events when doing a GET request (a known anitpattern),
but it will stay for now until it becomes clearer how the API should
look like in the end.