Add network interconnects mapping DB table
This commit adds the alembic DB migrations required to map special
devices like Transits or bodergateways (BGWs) to network/AZ
combinations.

There were some considerations of doing the bookkeeping for Transits
only via OpenStack ports (which then carry the required binding_host),
but there were two problems with this approach which made us decide that
we needed an extra table:

1. We'll have to query this table a lot and there is no proper index on
the ports table for the things we want to query. We could add this
index, but it would require modifying an OpenStack table, which is
generally frowned upon

2. We need to make sure there is only one Transit and one BGW per
network/AZ combination. Two concurrently running neutron-server
processes might both allocate a Tranit or BGW. With the table in this
commit and its unique constraints one of the processes will fail
and then find the already allocated device, i.e. we have locking between
these processes via the DB table.

Creating the migrations was a hard battle. For other Neutron subproject
it seems to work that their models are not included directly in env, for
us it didn't seem to work. Another problem is, that in its current form
neutron-db-manage does not allow --expand --autogenerate to work
togehter (mutually exclusive group), though without this we don't get
any autogenerated migrations that are not just empty templates. Patching
neutron-db-manage for this helped and now we have them.