[docs] Add Converged Cloud header / footer
This is a plugin for OpenStack, but it is not an OpenStack project. It
is (at this time) completely developed in-house with SAP. To make sure
people don't think this is an internal OpenStack project we're now
changing the template of the autogenerated docs, remove parts of the
OpenStack headers and replace this with our own logos.

Co-authored-by: Esther Schmitz <esther.schmitz@sap.com>