[docs] Remove CC license
To make sure we don't run in any licensing issues with internal company
policies we are - for now - removing the explicit CC licensing header.
The whole project is licensed under the Apache 2 license and I think
this applies also for the docs if not stated otherwise. As I'm a
software-dev with some minimal licensing knowledge and not an expert
(especially when it comes to corporate licensing) we're going with this
until some clarification has been made.