Use str representation for RoleEnum in config
To properly import/export the role enum we set `use_enum_values` to
true. This means when exporting the config the role will be presented as
a string (and not an unusable python class). In the end we might want to
use this for all pydantic classes, but for now we'll just use it with
Hostgroups where we have an Enum - for testing.