Implement Transit/BGW network allocation lifecycle
This commit implements Transit and BGW allocation for networks.

Transits and BGWs are allocated on network create in
create_network_postcommit. They are allocated for all AZs the driver
knows about. This decision (as opposed to "allocate devices per AZ
whenever a network enters/leaves an AZ) was made to reduce code
complexity, allocation complexity and the potential for race conditions
(concurrent portbindings for an enter/leave AZ condition). This comes at
the cost of having more networks on our Transits/BGWs. Note, that this
does not include AZ-aware networks, as these will not have BGWs
allocated to them, though this is not implemented yet (hence the FIXMEs
in the code).

A "global_config" section has been added to the driver config. This
section currently only carries the regional ASN, but will be used for
others, especially when BorderLeaf VRF configuration comes into the
picture.

Transits/BGWs have no OpenStack ports associated with them, therefore it
is forbidden for the user to bind ports for these themselves. In the
future there will be an API to sync config for these devices, but this
has not been implemented yet.