Implement AZ awareness / AZ-local networks
This driver can now handle AZ-local. This means that a network created
with an AZ hint will be handled as AZ-local. No BGWs will be scheduled,
as no inter-AZ traffic is allowed or required and no transits outside of
the network's AZ will be allocated.

Ports which map an a switchgroup that is in an AZ not matching with the
network's AZ hint will not be bound and an exception will be thrown.

A network can have a maximum of one AZ, this is enforced in
create_network_precommit().