Pin main to Neutron Ussuri
The first region where we're going to test the driver will be on Neutron
Ussuri. Upcoming changes in current Neutron main are incompatible with
this driver's testing framework. At least this commit started to make
things incompatible and requires us to branch off:

0e1ec521180bebe4a01298bc8cdec765cc1ec796

So for now we're going to stick with Ussuri until our testbed has been
updated. For production we need _at least_ Neutron Yoga, as this release
has some patches to prevent segment allocation race conditions.