Implement agent state reporting
To have an entry in the "openstack network agent list" aka registering
our own agent we need to report state. Originally we thought we could
use Service.report_state, but this does not work without subclassing it
(it creates a looping call that calls an empty method).

So to properly register ourselves we now take the way of the DHCP agent
(or the OVS agent or a couple of other neutron pieces) and implement our
own _report_state() method.

Having agent_type as an argument to Service.create() never worked by the
way, hence we removed this.