Implement Network/Switch/SwitchGroup Sync API
With this commit we can now inspect and sync networks, switches and
switchgroups.

Networks can't be listed (it would simply list all available network
ids), but can be shown by id - this will include some extra info by the
driver (binding hosts, interconnects). There are two extra methods
available: sync to sync the network (will be a config "merge") and
ensure_interconnects, which will allocate and configure interconnects
to said network. Both are available via HTTP PUT.

Switches can be listed and shown. Adding ?device_info=1 to the URL will
also fetch uptime and the current firmware version from the switch. Sync
is available, but will replace the config by fetching all physnets
scheduled onto this switch. InfraNetworks will be configured as well. To
configure only infranetworks sync_infra_networks is available.

SwitchGroup operations are implemented by searching for the given
switchgroup and then calling the SwitchesController for each switch in
the switchgroup. For operations on the switchgroup this means one RPC
call is done sequentially for each switch. This has some potential for
improvement, but for now this was the easiest way to implement it. When
this becomes a painpoint we can fix it.

"diff" is not yet implemented. It will be implemented together with
switch config fetching.

References to the CCDbPlugin have been replaced with the FabricPlugin.
The FabricPlugin acts very much like the CCDbPlugin, but it adds some
extra non-db methods. This was done because I needed a place to put code
which can be accessed both by the Ml2MechanismDriver and the API, but I
still wanted to keep the CCDbPlugin mostly about DB methods.  The fabric
plugin will most likely be used by the config syncloop as well.