Add run() method to agent
We copied part of the agent's _report_state() method out of neutron. The
code expects a run() method to call for actions to be done on first
agent start, but at the moment we don't have anything we would want to
call there. To allow us to fill this method later on (e.g. with a sync
loop) I just added this method with a simple log message. For future
use.