Always use vlan_pool as physnet, never the host
In this case this only happened for interconnects. We used a host where
we should have used the vlan_pool / physnet instead. The tests didn't
catch this, because the config fixtures were created with the same name
for binding_host and switchgroup name (i.e vlan pool). This commit fixes
the bug and uses different names for host/vlan pool (at least in the
interconnects tests) so we don't make this mistake again.