Set RD on non-bgws, change rt schema
BGP vlan instances need an RD by default, so this is now set to
{switch_asn}:{vni}.

The RT format was different between instances, which does not work - it
needs to be unified (which was clear from the beginning but somehow not
implemented this way). For now we go with {region_asn}:{vni}, but this
format might change and include the AZ in the future.