Handle broken bindings on port delete/update
In delete_port_postcommit and update_port_postcommit we make decisions
based on the binding levels given by the port (or the original port in
the update case). These levels might not be present (e.g. for an unbound
port) or they might straight out be broken (only bound on one level).
Now we only process ports, which have at least two binding levels, as
these are ports that are bound and expected to have config on the
fabric.