Add vlan ranges to driver config
GlobalConfig now has a setting to define the default vlan ranges. These
can be overridden on a per-switchgroup basis.

This setting is not used inside the driver code itself yet, but can be
used by external tools (e.g. helm-charts) to generate the
network_vlan_ranges setting for OpenStack's VlanTypeDriver.