Switch to stable/ussuri-m3 as curr default branch
In sync with the tradition of our other drivers we will have one branch
for each Neutron release we're currently maintaining or have maintained
in the format of stabel/$release-m3. Whatever we are currently running
in prod is the default branch on GitHub. We can currently do it this way
as we're the only user of this driver and do not need to maintain
multiple releases, except when we do a migration ourselves.

So, in short, this is what we're doing now, until a better release /
branch management model is needed.