Explicitly list AZs in driver config
In the future each AZ will need a number and a suffix. As we don't want
to require a specific AZ format, we allow this to be configured. This
also allows us a somewhat easier AZ handling, as we don't need to
iterate over all switchgroups and could also define AZs for which we
don't have a switchgroup.

Adaption of the config generator for this is not part of this commit.