Explicitly add RTs to agent message format
We now put the Route Targets that should be exported/imported directly
into the agent message / SwitchConfig format. This means that we'll have
less business logic inside the agent and more in the neutron-server part
of the driver (which is a good thing, as we will have multiple agents
and therefore will not need to duplicate business logic). It will also
help when we implement the config diff functionality, as we can then
just fill the SwitchConfig object with what's on the switch and diff it
with the "should" version from OpenStack.