Pull InfraNetworks from netbox and change driver config
Pulling InfraNetworks from netbox using vlans assigned to the server
facing ports, as well as the SVIs modelled on the TORs for the anycast
gateteway.
Also made VNI an mandatory field, as we potentially want to stretch
every VLAN over TOR borders.
VRF is also mandatory if the network is L3 enabled.