Add RPC call to get config from switches
We now can grab the config from a switch (or a number of switches). The
config will be put into a SwitchConfigUpdate object for transport and
vendor independency. This means that config snippets that are vendor
specific and/or are implicitly generated (for example the "redistribute
learned" flag in the EOS case) are not shown.

With this functionality we are one step closer to proper config diffing
capabilities.