Remove old Vxlan maps on config update
When introducing new vni-vlan mappings on switch config update
(operation add) we need to look onto the device if there is already an
old mapping for either the vni or vlan id that we are referencing. These
need to be removed in order for the new mappings to be properly applied
to the switch. Sadly, this means that we have to actually do a lookup on
the switch each time we add one or more vni/vlan mappings, but that is
just the world we life in. On add operation we log a warning if there
are stale mappings, as this generally should not happen that often (only
when "things break" or devices are unreachable).