Add aggregate attribute to InfraNetworks
We want to aggregate as much as possible. L3 enabled InfraNetworks for
us are usually supernetted on a TOR level to a single network. We could
calculate the aggregate if all networks in all InfraNetworks associated
with a hostgroup would result in a full supernet but they do not.

In order to not hardcode the aggregate calculation logic into the code
and stay flexible, we feed an optional aggregate list for each
InfraNetwork. This indicates which aggregate the InfraNetwork will be
part of.

In turn, this aggregation list is for us generated out of the netbox
parent prefix.