Replace eAPI with grpc/GNMI for Arista EOS agent
The Arista eAPI has two options to get information about the switch:
show commands (which show the state of the device) and a dump of the
running-config (which dumps the whole running-config). We need config
state, not device-state, to properly reconfigure the device and we do
not want to rely on device state, as this state sometimes depends on
external factors (e.g. a shut device might not show its config via
exposed state). We also don't want to write another running-config
parser.

Therefore, we decided to go with GNMI. This is based on grpc and YANG
models. Arista EOS uses openconfig YANG models, but augments them at
some places to put in some extra infos (that we also do need).

Some parts of the config are not yet properly modeled (at this point in
time this is everything we need for the BGW functionality), so there we
need to access the sysdb via GNMI to get the info off the device (not
properly implemented in our driver yet) and need to use cli commands to
configure the BGW functionality.