Use switchgroup id as RD for BGP vlans
The BGP class of agent messages now carries a `switchgroup_id`. This
attribute is used by add_vlan() (and later add_vrf()) to generate the RD
for each BGP vlan stanza. It will not be filled when pulling config from
a switch and will probably need to be deleted when comparing config.

We also now change the RD, as just implied. The RD of BGP vlans is now
in the format of "{switchgroup_id}:{vni}", as the old format (prepending
the region asn in the format of "{region_asn}.{switchgroup_id}:{vni}"
was not available with Arista GNMI - they just break / can't parse these
Type 2 RDs. So, this is the way it shall be now.