Properly replace trunk vlans on delete
When we delete a vlan from an interface we need to do a replace with the
set of vlans that do not contain the vlan that we want to delete. This
needs an explicit replace on the trunk-vlans entity, as else we would
delete parts of the interface that we didn't plan on touching.