Change VLAN deletion order check
For VLAN-VNI mappings we check that the VNI is mapped to the right VLAN
(the one we want to delete), but before this commit this check was done
by iterating over all existing switch-VLANs and look if they exist in
the OpenStack config request. This leads to a lot of "VLAN not found on
switch" warnings, which is... wrong. Iterating over the OpenStack VLANs
and then looking on the switch is what was originally intented.