Fixes for dry-run mode
When running in dry-run mode, objects might or might not exist in netbox. If
they don't we use the dict that we'd elsewise use to create the object as a
mock. Unfortunately dicts are not hashable and also attribute access needs to
be treated a bit different. We address 2 occurence of these in this commit.