Use different RD/RT numbering scheme due to unsupported type 2 RDs
GNMI does not support type 2 RD according to RFC4362 sec. 4.2 event
though the models state they do. Until this is fixed, we decided to use
type 0 RD and RTs. With type 0 RDs, we have only 2 bytes for the
administrator field and hence need to cut some bytes. We do that by just
using the switchgroup id instead of the whole ASN. As the ASN was
constructed with `f'{regional_asn}.{switchgroup_id}'` we are not loosing
any information here.