Workaround for `network`-statements not present in OpenConfig
While Arista provided an _augment_ for aggregate addresses it does not
provide one for `network` statements. We hence need to employ a little
workaround where we are using redistribute maps and prefix lists to get
the necessary RTs on our routes based on their properties. This change
includes the workaround and a table to better show route properties and
their config consequence.