Add API call to create all port-channels
With GNMI on Arista EOS we cannot reference a Port-Channel that already
exists. As we don't want to split up our set() call in the agent and we
only would need to do this a single time for the first time provisioning
a switch we now offer an extra API call to create these Port-Channels on
all devices.

I really would have liked to do this via collection_methods or
collection_actions (comparable to member_actions) and there is also an
example for this in Neutron (see the tagging extension or the quota
extension), but I did not get this to work. Therefore I now do a
"switch-case" in update() of the switch controller.