Use collection actions in API
Originally we used a switch-case style approach in SwitchesController
update(), but there is a facility in the Neutron API, which can help
with having an action on the collection - the collection actions!
Originally this did not work, but with adding an extra / to the
path_prefix everything seems to be in order. This also doesn't seem to
impact other API calls, as all tests seem to run through just fine.

Co-Authored-By: Johannes Kulik <johannes.kulik@sap.com>