Support starting the agent with threading
This module bundles some functionality of oslo_service, neutron.service
and neutron_lib.rpc into a single class to be able to start the agent
with threading instead of eventlet.
Especially the signal-handling build deeply into the stack creates the
need to call all functionality on our own. There's still some room for
improvement in what the main thread does, as it's currently only
sleeping during the main work to wait for signals to arrive.