Switch agent to use threading instead of eventlet
With our agent's underlying grpc library being written in C and not
having a custom eventlet I/O handler, the whole process blocks for the
time of a request to the switch. This leads to timeouts in reporting
state to neutron-server and also in our liveness-checks.

To mitigate this, we want to switch from eventlet, which is based on
cooperatively giving up the CPU mainly via I/O handlers, to threading.
We've implemented a couple of helper modules to make the code as
compatible as possible during the switch.

Attention: While this commit switches the agent over to threading, we
still need follow-ups to prohibit multiple threads talking to the same
switch in parallel and thus race-conditions in configuration a switch.