Replace only managed vlan objects on config replace
It was decided that we do not own all the VLANs and connected objects
(vlan-vni mappings, evpn instances) on a device. Therefore, we cannot
just do a config replace and tell the switch of all the VLANs that we
want to have on the device. What we have to do now is to look at all the
VLANs on the device, see if they are still needed and if not AND are
managed by us we delete them - case by case. The range of managed VLANs
is pulled from the driver config - all vlan ranges + all infra net vlans
that we manage.