Only do one config update per switch
With the switch of the threading module we need to make sure only one
apply_config_update() runs at a time on a per switch basis.
apply_config_update() calls get() to the API several times, works on
this data and then does a set(). Things on the switch should not change
in the meantime.

Concurrent get()s (e.g. get_config()) might yield an inconsistent view,
but that would be okay, as this call is there to show the current state
of the device, not a necessarily consistent one.