Support unmanaged interfaces
An unmanaged interface is any interface we want to keep in the driver
config, but do not want to push any config for. This has been introduced
for transit interconnects, as we don't have the full vlan list to
properly control the allowed vlan list for now, but it can also be used
for other interfaces. Due to it being thought out as a transit feature I
decided to also model the config fixtures for the tests according to
this and make transit ports by default unmanaged.

I chose to use unmanaged as a config option in favor to just making all
transit interfaces unmanaged by default, because this is a Converged
Cloud design decision. The driver could very well be run with managed
transit interfaces (as opposed to the BGWs, where we just don't have any
interface to manage).