Support starting a backdoor shell
To replace oslo.service's backdoor, we use "manhole" to spawn a shell in
a thread. We often connect to the backdoor shell to debug problems
inside the context of the application, but we cannot use oslo.service
anymore when using threads as it's too tightly coupled to eventlet.
We're using the same config option backdoor_socket as oslo.service here,
which can be used to disable the backdoor shell by not setting it or
setting it to an empty value.