Read and replace EVPN route targets on BGWs
EVPN route targets (and also the EVPN route distinguisher) are somwhat
hard to read, as they are only present in the eos_native:Sysdb, not in
the augmented models. This format is somwehat different from what we
know and love and the values we want are only present as an integer
representation (the route distinguisher is only presnet in network byte
order, which was a really fun thing to find out, but "should have been
obvious" from the NBO in the name).

With this, we now can properly do a replace on BGWs, as we can clear out
stale route targets. This is possible, as we now have visibility on
what's actually on the device.

I'm really looking forward to when we have models for the BGW
functionality and can remove the cli-via-gnmi mode altogether.