Cache frequently used mappings in driver config
Some of the helper methods to find certain things in the config (e.g.
get a hostgroup by host, get a switchgroup by switch member name) are
called very often and while they are pretty fast, some of them are
called so often that this accumulates to quite a lot of time! To speed
this up a bit we now pre-cache some of these mappings and then just do a
dictionary access as - famous last words - the config does not change
while the driver is running.

We do a similar thing with get_vlan_pool_name(), where we cache the vlan
pool name on first call.