Allow Hostgroups to have "extra vlans"
"Extra vlans" are vlans, that should be present on an interface, but get
no vxlan-vlan mapping, no bgp entries and don't depend on a port binding
to be present. Similar to infra networks they will only be provisioned
on switch config replace or when replacing the whole switch config. They
will be realized as a simple extra trunk vlan in the "trunk allowed
vlans" list of an interface.