Always set RD, even if set via cli for BGWs
We need to always set the route-distinguisher. In the past we didn't do
this when we already configured the RD via cli in the case of BGWs, as I
thought this was already enough. If we set the RD only via cli with "rd
evpn domain all" and THEN do a replace on this evpn instance without a
RD being present in the config request the cli statement will change to
"rd evpn domain remote" and won't export routes learned from other BGWs.
As we want a working fabric, we now always add the RD to an evpn
instance config request.