Keep context in ThreadPoolExecutor
When submitting a function into the ThreadPoolExecutor, we don't run in
the same thread anymore and thus loose the thread-local context
oslo.logging uses for logging the request ID. To mitigate this, we now
fetch the local context before submitting the function into the
ThreadPoolExecutor and submit a wrapped function that sets the
thread-local context again. If we didn't have context to begin with, we
remove any currently-set thread-local context, so we don't log wrong
request IDs and don't allow leaking of a context into another function.