Modularize infra networks generation/add extra vlan support
There might be infra networks that are not BB specific, they are not
there yet, but they might come. Hence the VNI generation which very much
depends on the scope of an infra network was moved into a dedicated
method. Also the check and generation of L3 attributes of an infra
network was moved into another method to make the infra network
generator smaller. The generator now also returns extra vlans.