Sync switchgroup in one go via API
With most of the API calls of SwitchgroupsController we just call the
SwitchesController for each member of the switchgroup. If we don't
expect a call to take a long time this might be okay, but with the
switch sync taking something between 15-40s for some devices we want to
parallelize these requests. Therefore we now generate the
SwitchConfigUpdateList once for the switchgroup (which is essentially
what we did before) let the SwitchesController clean out the switch it
doesn't need to sync. With this SwitchgroupsController can just use this
functionality and sync a switchgroup with a single RPC call.