Switch to distinct read and write ThreadPools
We don't want the writes to starve out the reads just because they are
slower and clog up the queue. We also want better statistics on the
average runtime for writes - which take a lock and thus block other
threads currently executed in the ThreadPool from finishing.

Therefore, we introduce a second ThreadPool and use one for reading and
one for writing.

This commit does the appropriate changes to SwitchBase, the
backdoor-locals, the agent and our custom GMR reporting.