Do not redirect stderr in backdoor
We don't tell manhole to redirect stderr anymore, because it would make
the GuruMeditationReport print only into the manhole thread and not into
the logs anymore. This defeats part of the purpose of the GMR: that we
still get a state from the process, even if the backdoor doesn't work
anymore.
Additionally, redirecting stderr didn't have the wanted effect of
showing us tracebacks/exceptions of code we run in the backdoor. We need
to investigate further how we could make that work.