Add periodic task to persist switch configs
We now have a periodic task that takes care of persisting all the switch
configs. By default it is run every 15 minutes. For EOS it will simply
do a "copy running-config startup-config".

In the future EOS might automatically persist its config, then we can
just override the method in the EOS switch agent. NXOS - when it will
come - will still need it.