Remove agent check / ping_back
This was one of the first checks implemented in the driver RPC API and
used to test the communication between driver and agent. With the
upcoming switch config sync loop we have more than enough methods to use
the API.

Also, the ping_back imported the CCFabricDriverRPCClient, which resulted
in an import cycle. This could have been solved otherwise, but as we
don't need this functionality, I'm just removing it.