Add sync loop to periodically sync all switches
We now have a sync loop in the agent that syncs every switch - in
parallel - with the switch config it should currently have. By default
it runs every 15 minutes.

This config is pulled right when the task starts and not before we put
it into the writer queue, because we want to make sure that we have the
most recent config once the task is actually starting. The config is
pulled via the driver's RPC API and uses the same method that is also
used by the sync API as part of the driver's OpenStack API /
fabricoperations.