API for managing syncloops on agents
Sometimes when doing certain chainges, like changing out configs,
exchanging switches or doing some DB work we need to pause the
syncloops to make sure we don't sync garbage onto our infrastructure. To
make this task easy we now have an API to enable/disable the switch
syncloop on a per agent basis.

Doing a GET on this resource shows the time of the last successful sync
for every switch. Getting a None means this switch has never been
successfully synced, either because of a problem or because there is no
config whatsoever available for this device. The latter "should" only be
a problem for our QA environments.