Use `external` keyword to differentiate from BGP aggregates
In our network we will announce networks to other regions or peering
partners based off a certain community. We use route-maps and prefix
lists to employ a this community indicating that a network should be
externally announced.

Earlier we referred to these networks bluntly as _aggregates_ as we were
always announcing aggregates externally. Yet that terminology was
confusing as a network may not be an aggregate but still should be externally
announced and then ended up in a prefix list that was called something
with _aggregate_. Hence, we change our naming and refer to them as
externally announced networks and name the connected prefix list
like `PL-CC-CLOUD02-A-EXTERNAL`.