[docs] Reorganized docs, created section that defines config generation
Rearranged a lot. Basically created a section for the config to define
what the config objects to. Carved that out from the now non-existing
config-input page and renamed the remainder to config-gen. That page now
focuses more around what the config-generator expects and why something
might not show up. Still work to be done here for stpod, apod, netpod.

I also removed all sample code with a reference to the modeller.