Fix l3 bgp lifecycle
VRF Networks (i.e. OpenStack subnet CIDRs) now are configured and
deconfigured together with the segment joining and leaving a physnet (in
most cases == switch). This makes sense, as only one subnet inside of a
VRF should be able to have a certain CIDR (if this changes, we have a
severely broken config and most likely other problems as well).
Aggregates on the other hand might have a different lifecycle, as
multiple networks in the same subnetpool might share some aggregates.
We now could do some DB queries for this and try to find out what the
situation is for a certain subnetpool and figure out what to add or
delete... But at the moment I want to get this out of the door,
therefore we choose the simple approach: Let it be. The switch syncloop
will take care of this every 15 minutes, so we have bogus aggregates for
a maximum time of 15 minutes and they should not break anything for us.
If we find out they do then we'll have to revisit this.

The original idea of the l3 lifecycle somehow assumed that one network
could be on the same physnet multiple times, which is not possible. I
probably got this mixed up with the aggregate lifecycle.